style: Moving the updating DTOs to its own directory
